Break out sse into a compile unit
Verify similar to beb96b0 Break out websockets

Break out the SSE functions into a separate compile unit.  This allows
the SSE sockets in beast to be compiled separately, which significantly
reduces the overall compile time by a few seconds.  Code is identical
with the exceptions of minor header definitions to convert header-only
to compile unit.
